Add revive modules for Farooq's Revive
The *Instant Revive* module will immediately revive the selected player
when Farooq's Revive script is active. It still takes seconds for the
player to be up because of Farooq's Revive script.

The *Immersive Revive* module sends a selected AI unit to the nearest
down player to revive him. This gives a more realistic feeling.