Add pickup waypoint
This waypoint first lands a helicopter and then waits for players to enter it.
If no new player enters for 10 seconds the next waypoint will be taken.

This is based on fn_pwLand.sqf but only handles one heli in the group. It's
unclear to me how there could be multiple.